Conflicts:
	app/controllers/points_controller.rb
	app/controllers/solutions_controller.rb
	app/controllers/submissions_controller.rb
	app/models/course.rb
	app/views/exercises/show.html.erb
	app/views/submissions/_submission_details.html.erb
	config/routes.rb
	spec/controllers/points_controller_spec.rb

+++ app/models/course.rb
@@ -11,15 +11,26 @@ class Course < ActiveRecord::Base
   validates :name,
             presence: true,
             uniqueness: true,
-            length: { within: 1..40 },
             format: {
               without: / /,
               message: 'should not contain white spaces'
             }
+  validates :title,
+            presence: true,
+            length: { within: 1..40 }
+  validates :description, length: { maximum: 512 }
+  validate :check_name_length
 
+  # If made from template, make sure cache_version is not out of sync.
+  before_save :set_cache_version
+  before_validation :save_template
   validates :source_url, presence: true
-  validate :check_source_backend
-  after_initialize :set_default_source_backend
+  #validates :custom_points_url,
+  #          format: {
+  #            with: /(\Ahttps?:\/\/|\A\z|^$)/,
+  #            message: 'should begin with http:// or https://'
+  #          }
+  validate :check_external_scoreboard_url
 
   has_many :exercises, dependent: :delete_all
   has_many :submissions, dependent: :delete_all
@@ -35,6 +46,7 @@ class Course < ActiveRecord::Base
   has_many :assistantships, dependent: :destroy
   has_many :assistants, through: :assistantships, source: :user
 
+  belongs_to :course_template
   belongs_to :organization
 
   scope :with_certificates_for, ->(user) { select { |c| c.visible_to?(user) && c.certificate_downloadable_for?(user) } }
@@ -55,11 +67,49 @@ class Course < ActiveRecord::Base
 
   scope :ongoing, -> { where(['hide_after IS NULL OR hide_after > ?', Time.now]) }
   scope :expired, -> { where(['hide_after IS NOT NULL AND hide_after <= ?', Time.now]) }
-  scope :assisted_courses, ->(user) { joins(:assistantships).where(assistantships: { user_id: user.id }) }
-  scope :participated_courses, ->(user) do
+  scope :assisted_courses, ->(user, organization) do
+    joins(:assistantships)
+      .where(assistantships: { user_id: user.id })
+      .where(organization_id: organization.id)
+  end
+  scope :participated_courses, ->(user, organization) do
     joins(:awarded_points)
       .where(awarded_points: { user_id: user.id })
+      .where(organization_id: organization.id)
       .group('courses.id')
+  end
+
+  def self.new_from_template(course_template)
+    Course.new(name: course_template.name,
+               title: course_template.title,
+               description: course_template.description,
+               material_url: course_template.material_url,
+               cache_version: course_template.cache_version,
+               course_template: course_template)
+  end
+
+  def git_branch
+    course_template_obj.git_branch
+  end
+
+  def source_url
+    course_template_obj.source_url
+  end
+
+  def source_backend
+    course_template_obj.source_backend
+  end
+
+  def git_branch=(branch)
+    course_template_obj.git_branch = branch
+  end
+
+  def source_url=(url)
+    course_template_obj.source_url = url
+  end
+
+  def source_backend=(backend)
+    course_template_obj.source_backend = backend
   end
 
   def visible_to?(user)
@@ -102,7 +152,6 @@ class Course < ActiveRecord::Base
     self.hidden = !!new_options['hidden']
     self.spreadsheet_key = new_options['spreadsheet_key']
 
-    self.description = new_options['description']
     self.paste_visibility = new_options['paste_visibility']
     if !new_options['locked_exercise_points_visible'].nil?
       self.locked_exercise_points_visible = new_options['locked_exercise_points_visible']
@@ -138,8 +187,12 @@ class Course < ActiveRecord::Base
     "#{FileStore.root}/course"
   end
 
+  def increment_cache_version
+    course_template_obj.increment_cache_version
+  end
+
   def cache_path
-    "#{Course.cache_root}/#{name}-#{cache_version}"
+    course_template_obj.cache_path
   end
 
   # Holds a clone of the course repository
@@ -214,12 +267,12 @@ class Course < ActiveRecord::Base
     @groups = nil
   end
 
-  def refresh
-    CourseRefresher.new.refresh_course(self)
+  def refresh(options = {})
+    CourseRefresher.new.refresh_course(self, options)
   end
 
   def delete_cache
-    FileUtils.rm_rf cache_path
+    FileUtils.rm_rf cache_path if custom?
   end
 
   def self.valid_source_backends
@@ -269,13 +322,18 @@ class Course < ActiveRecord::Base
   def submissions_to_review
     submissions.where('(requests_review OR requires_review) AND NOT reviewed AND NOT newer_submission_reviewed AND NOT review_dismissed')
   end
-
+  
   def certificate_downloadable_for?(user)
     user.administrator? || (
       !user.guest? &&
        certificate_downloadable &&
        (certificate_unlock_spec.nil? ||
         UnlockSpec.new(self, ActiveSupport::JSON.decode(certificate_unlock_spec)).permits_unlock_for?(user)))
+  end
+
+  def toggle_submission_result_visiblity
+    self.hide_submission_results = !self.hide_submission_results
+    save!
   end
 
   # Returns a hash of exercise group => {
@@ -322,8 +380,8 @@ class Course < ActiveRecord::Base
     result
   end
 
-  def initially_refreshed?
-    refreshed_at.nil?
+  def refreshed?
+    !refreshed_at.nil?
   end
 
   def taught_by?(user)
@@ -342,15 +400,81 @@ class Course < ActiveRecord::Base
     name
   end
 
+  def material_url=(material)
+    return super('') if material.blank?
+    unless material =~ /^https?:\/\//
+      return super("http://#{material}")
+    end
+    super(material)
+  end
+
+  def custom?
+    course_template_obj.dummy?
+  end
+
+  def external_scoreboard_url=(url)
+    return super("http://#{url}") unless url =~ /^(https?:\/\/|$)/
+    super(url)
+  end
+
+  def contains_unlock_deadlines?
+    exercise_groups.any? { |group| group.contains_unlock_deadlines?}
+  end
+
+  def has_external_scoreboard_url?
+    !external_scoreboard_url.blank?
+  end
+
+  def parsed_external_scoreboard_url(organization, course, user)
+    external_scoreboard_url % { user: user.username, course: course.id.to_s, org: organization.slug }
+  end
+
   private
 
-  def check_source_backend
-    unless Course.valid_source_backends.include?(source_backend)
-      errors.add(:source_backend, 'must be one of [' + Course.valid_source_backends.join(', ') + ']')
+  def set_cache_version
+    self.cache_version = course_template_obj.cache_version
+  end
+
+  def save_template
+    course_template_obj.save!
+  rescue => e
+    course_template_obj.errors.full_messages.each do |msg|
+      errors.add(:base, msg + e.message)
     end
   end
 
-  def set_default_source_backend
-    self.source_backend ||= Course.default_source_backend
+  def course_template_obj
+    self.course_template ||= CourseTemplate.new_dummy(self)
+  end
+
+  def check_name_length
+    # If name starts with organization slug (org-course1), then check that
+    # the actual name (course1) is within range (for backward compatibility).
+    if name.start_with?("#{organization.slug}-")
+      test_range = name_range_with_slug
+    else
+      test_range = name_range
+    end
+
+    unless test_range.include?(name.length)
+      errors.add(:name, "must be between #{name_range} characters")
+    end
+  end
+
+  def name_range
+    1..40
+  end
+
+  def name_range_with_slug
+    add_length = organization.slug.length + 1
+    (name_range.first + add_length)..(name_range.last + add_length)
+  end
+
+  def check_external_scoreboard_url
+    begin
+      external_scoreboard_url % { user: '', course: '', org: '' } unless external_scoreboard_url.blank?
+    rescue
+      errors.add(:external_scoreboard_url, 'contains invalid keys')
+    end
   end
 end
